# services/gateway-svc/api/rest_oapi/openapi.yaml
openapi: 3.0.3
info:
  title: gateway-svc API
  version: 0.1.0
  description: |
    Edge REST API for UmrohOS. gateway-svc fronts every backend service and is
    the only outward-facing HTTP surface clients (web app, mobile app, CLI)
    should call.

    Scaffold scope: 12 endpoints —

      - 2 own probes: /system/live, /system/ready
      - 10 per-backend liveness proxies: /v1/<svc>/system/live, one per
        backend in the umrohos service map. Used by the web app's status page.

    Real per-route proxies (auth, packages, bookings, ...) land alongside
    each backend's first feature work.
servers:
  - url: /
    description: Base path
paths:
...
  /v1/packages:
    get:
      tags: [Catalog]
      summary: List active packages
      operationId: listPackages
      security: []
      parameters:
        - name: kind
          in: query
          required: false
          schema: { $ref: "#/components/schemas/PackageKind" }
        - name: airline_code
          in: query
          required: false
          schema: { type: string }
        - name: hotel_id
          in: query
          required: false
          schema: { type: string }
        - name: departure_from
          in: query
          required: false
          schema: { type: string, format: date }
        - name: departure_to
          in: query
          required: false
          schema: { type: string, format: date }
        - name: cursor
          in: query
          required: false
          schema: { type: string }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ListPackagesResponse" }
        "400":
          description: Bad request (invalid_query_param / invalid_cursor)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
        "500":
          description: Internal error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
        "502":
          description: catalog-svc unreachable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
  /v1/packages/{id}:
    get:
      tags: [Catalog]
      summary: Active package detail
      operationId: getPackageById
      security: []
      parameters:
        - name: id
          in: path
          required: true
          description: Package ULID, `pkg_`-prefixed.
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GetPackageResponse" }
        "404":
          description: Package not found or not publicly visible.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
        "500":
          description: Internal error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
        "502":
          description: catalog-svc unreachable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
  /v1/package-departures/{id}:
    get:
      tags: [Catalog]
      summary: Active package-departure detail
      operationId: getPackageDepartureById
      security: []
      parameters:
        - name: id
          in: path
          required: true
          description: Departure ULID, `dep_`-prefixed.
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GetDepartureResponse" }
        "404":
          description: Departure not found or not publicly visible.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
        "500":
          description: Internal error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
        "502":
          description: catalog-svc unreachable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }

  # ---------------------------------------------------------------------------
  # Staff catalog write routes (BL-CAT-014 / S1-E-07)
  # All require Bearer token + catalog.package.manage permission.
  # Response shapes are the same read models (PackageDetail, DepartureDetail)
  # so the frontend can reuse its existing display components.
  # ---------------------------------------------------------------------------
  /v1/packages/create:
    # POST /v1/packages is handled inline under /v1/packages above, but since
    # oapi-codegen merges methods on the same path, we declare it here as a
    # separate path alias so it appears in the spec for documentation purposes.
    # The actual route registration is done manually in cmd/server.go.
    post:
      tags: [Catalog-Staff]
      summary: Create a catalog package (draft)
      operationId: createPackage
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePackageBody" }
      responses:
        "201":
          description: Package created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GetPackageResponse" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
        "403":
          description: Missing catalog.package.manage permission
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
        "500":
          description: Internal error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
...
  /v1/packages/{id}/departures/create:
    post:
      tags: [Catalog-Staff]
      summary: Create a departure under a package
      operationId: createDeparture
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Package ULID
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateDepartureBody" }
      responses:
        "201":
          description: Departure created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GetDepartureResponse" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
        "403":
          description: Missing catalog.package.manage permission
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
        "404":
          description: Package not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: PASETO
      description: |
        Single-point edge authentication per ADR 0009 / F1-W7. Clients send
        `Authorization: Bearer <access_token>` on any protected route.
        gateway-svc validates the token once against `iam-svc.ValidateToken`
        (gRPC); backends do NOT re-validate. Failure modes: missing/malformed
        → 401 UNAUTHORIZED; iam-svc rejects → 401 UNAUTHORIZED; iam-svc
        unreachable → 502 SERVICE_UNAVAILABLE (fail-closed).

        No route is currently tagged with this scheme — first consumer lands
        with BL-IAM-018 / S1-E-12 when iam auth REST routes move to gateway.
  schemas:
...
    CatalogErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: |
                snake_case catalog error code (e.g. `package_not_found`,
                `departure_not_found`, `invalid_query_param`,
                `invalid_cursor`, `internal_error`, `service_unavailable`).
              example: package_not_found
            message:
              type: string
              description: Human-readable message in Bahasa Indonesia (Q003).
              example: paket tidak ditemukan
            trace_id:
              type: string
              description: OTel span ID (hex) for correlation with Tempo.

    # ── Catalog: money / enum helpers ──────────────────────────────────
    Money:
      type: object
      required: [list_amount, list_currency, settlement_currency]
      properties:
        list_amount:
          type: integer
          format: int64
          example: 38500000
        list_currency:
          type: string
          example: IDR
        settlement_currency:
          type: string
          enum: [IDR]
          example: IDR
    PackageKind:
      type: string
      enum: [umrah_reguler, umrah_plus, hajj_furoda, hajj_khusus, badal, financial, retail]
    RoomType:
      type: string
      enum: [double, triple, quad]
    DepartureStatus:
      type: string
      enum: [open, closed]
    OperatorKind:
      type: string
      enum: [airline, rail, bus]
    VendorReadinessState:
      type: string
      enum: [not_started, in_progress, ready, blocked]

    # ── Catalog: master-data refs ──────────────────────────────────────
    HotelRef:
      type: object
      required: [id, name, city, star_rating, walking_distance_m]
      properties:
        id: { type: string }
        name: { type: string }
        city: { type: string }
        star_rating: { type: integer, minimum: 0, maximum: 5 }
        walking_distance_m: { type: integer, minimum: 0 }
    AirlineRef:
      type: object
      required: [id, code, name, operator_kind]
      properties:
        id: { type: string }
        code: { type: string }
        name: { type: string }
        operator_kind: { $ref: "#/components/schemas/OperatorKind" }
    MuthawwifRef:
      type: object
      required: [id, name, portrait_url]
      properties:
        id: { type: string }
        name: { type: string }
        portrait_url: { type: string }
    AddonRef:
      type: object
      required: [id, name, list_amount, list_currency, settlement_currency]
      properties:
        id: { type: string }
        name: { type: string }
        list_amount: { type: integer, format: int64 }
        list_currency: { type: string, example: IDR }
        settlement_currency: { type: string, enum: [IDR], example: IDR }
    ItineraryDay:
      type: object
      required: [day, title, description]
      properties:
        day: { type: integer, minimum: 1 }
        title: { type: string }
        description: { type: string }
        photo_url: { type: string }
    Itinerary:
      type: object
      required: [id, days, public_url]
      properties:
        id: { type: string }
        days:
          type: array
          items: { $ref: "#/components/schemas/ItineraryDay" }
        public_url: { type: string }

    # ── Catalog: departure-detail shapes ───────────────────────────────
    PackagePricing:
      type: object
      required: [room_type, list_amount, list_currency, settlement_currency]
      properties:
        room_type: { $ref: "#/components/schemas/RoomType" }
        list_amount: { type: integer, format: int64 }
        list_currency: { type: string, example: IDR }
        settlement_currency: { type: string, enum: [IDR], example: IDR }
    VendorReadiness:
      type: object
      required: [ticket, hotel, visa]
      properties:
        ticket: { $ref: "#/components/schemas/VendorReadinessState" }
        hotel: { $ref: "#/components/schemas/VendorReadinessState" }
        visa: { $ref: "#/components/schemas/VendorReadinessState" }
    DepartureDetail:
      type: object
      required:
        - id
        - package_id
        - departure_date
        - return_date
        - total_seats
        - remaining_seats
        - status
        - pricing
        - vendor_readiness
      properties:
        id: { type: string }
        package_id: { type: string }
        departure_date: { type: string, format: date }
        return_date: { type: string, format: date }
        total_seats: { type: integer, minimum: 0 }
        remaining_seats: { type: integer, minimum: 0 }
        status: { $ref: "#/components/schemas/DepartureStatus" }
        pricing:
          type: array
          items: { $ref: "#/components/schemas/PackagePricing" }
        vendor_readiness: { $ref: "#/components/schemas/VendorReadiness" }
    GetDepartureResponse:
      type: object
      required: [departure]
      properties:
        departure: { $ref: "#/components/schemas/DepartureDetail" }

    # ── Catalog: package-specific ──────────────────────────────────────
    NextDeparture:
      type: object
      required: [id, departure_date, return_date, remaining_seats]
      properties:
        id: { type: string }
        departure_date: { type: string, format: date }
        return_date: { type: string, format: date }
        remaining_seats: { type: integer, minimum: 0 }
    DepartureSummary:
      type: object
      required: [id, departure_date, return_date, remaining_seats, status]
      properties:
        id: { type: string }
        departure_date: { type: string, format: date }
        return_date: { type: string, format: date }
        remaining_seats: { type: integer, minimum: 0 }
        status: { $ref: "#/components/schemas/DepartureStatus" }
    PackageListItem:
      type: object
      required: [id, kind, name, description, cover_photo_url, starting_price]
      properties:
        id: { type: string }
        kind: { $ref: "#/components/schemas/PackageKind" }
        name: { type: string }
        description: { type: string }
        cover_photo_url: { type: string }
        starting_price: { $ref: "#/components/schemas/Money" }
        next_departure:
          allOf:
            - $ref: "#/components/schemas/NextDeparture"
          nullable: true
    PackageDetail:
      type: object
      required:
        - id
        - kind
        - name
        - description
        - highlights
        - cover_photo_url
        - hotels
        - add_ons
        - departures
      properties:
        id: { type: string }
        kind: { $ref: "#/components/schemas/PackageKind" }
        name: { type: string }
        description: { type: string }
        highlights:
          type: array
          items: { type: string }
        cover_photo_url: { type: string }
        itinerary:
          allOf:
            - $ref: "#/components/schemas/Itinerary"
          nullable: true
        hotels:
          type: array
          items: { $ref: "#/components/schemas/HotelRef" }
        airline:
          allOf:
            - $ref: "#/components/schemas/AirlineRef"
          nullable: true
        muthawwif:
          allOf:
            - $ref: "#/components/schemas/MuthawwifRef"
          nullable: true
        add_ons:
          type: array
          items: { $ref: "#/components/schemas/AddonRef" }
        departures:
          type: array
          items: { $ref: "#/components/schemas/DepartureSummary" }
    PageMeta:
      type: object
      required: [has_more]
      properties:
        next_cursor:
          type: string
          nullable: true
        has_more: { type: boolean }
    ListPackagesResponse:
      type: object
      required: [packages, page]
      properties:
        packages:
          type: array
          items: { $ref: "#/components/schemas/PackageListItem" }
        page: { $ref: "#/components/schemas/PageMeta" }
    GetPackageResponse:
      type: object
      required: [package]
      properties:
        package: { $ref: "#/components/schemas/PackageDetail" }

    # ── IAM auth schemas (BL-IAM-018 / S1-E-12) ─────────────────────────
    # All auth routes proxied from iam-svc REST → gateway REST → iam-svc gRPC.
...
    PricingInputBody:
      type: object
      required: [room_type, list_amount, list_currency, settlement_currency]
      properties:
        room_type:           { type: string, enum: [double, triple, quad] }
        list_amount:         { type: integer, format: int64, minimum: 1 }
        list_currency:       { type: string, enum: [IDR, USD] }
        settlement_currency: { type: string, enum: [IDR], default: IDR }
    CreatePackageBody:
      type: object
      required: [kind, name]
      properties:
        kind:            { $ref: "#/components/schemas/PackageKind" }
        name:            { type: string, maxLength: 200 }
        description:     { type: string }
        cover_photo_url: { type: string }
        highlights:      { type: array, items: { type: string } }
        itinerary_id:    { type: string }
        airline_id:      { type: string }
        muthawwif_id:    { type: string }
        hotel_ids:       { type: array, items: { type: string } }
        addon_ids:       { type: array, items: { type: string } }
        status:          { type: string, enum: [draft, active], default: draft }
...
    CreateDepartureBody:
      type: object
      required: [departure_date, return_date, total_seats]
      properties:
        departure_date: { type: string, format: date, description: "ISO YYYY-MM-DD" }
        return_date:    { type: string, format: date, description: "ISO YYYY-MM-DD" }
        total_seats:    { type: integer, minimum: 1 }
        status:         { type: string, enum: [open, closed], default: open }
        pricing:        { type: array, items: { $ref: "#/components/schemas/PricingInputBody" } }
...